Proxy objects.
Utilizing a class specifically for proxy objects.  They are ways to allow access to the control and value within it.

Adding a specific class for getter setter object.  Make sure to add tests specificially for it.  It supports default operators including streaming.

Updating documentation to clarify naming convention.

Adding proxy to example.