issue #10: Adding more targets to CI.
Fixing up to have wxWidgets be fetch content, so we get what we need from source.
Fixing up the type errors that we see on linux.  Refining the way we specify the version to build.
only run tests on Mac because we're doing weird things with wxWidgets.
Adding Debug and Release.
Fixing up std::nullopt.
Updating releases notes.
disable sanitize on MSVC