chore: try to find wxWidgets before auto pulling
Try to find wxWidgets with find_package before using `FetchContent`. If
you already have wxWidgets install, either from a binary, a package
manager or a OS distribution, you can save the extra build time of
pulling wxWidgets and building it as part of the main project when
building tests and/or the example.