Issue #219: We should have customization points for controllers to allow for testability
Making the way that we use Parent to be more generic, as it essentially is the thing that can create new controllers.
Using customaization functions and CPO to allow exentsion to types that conform to the API.  This allows something like TestProvider.
Moving tests to not rely on creating a wxWidget frame, which has major problems on windows, and on recent versions of wxWidgets.